requirements.txt
Flask==0.10.1
Pillow==2.7.0
numpy==1.9.1
